admin_api/pyproject.toml
[tool.poetry]
name = "admin-api"
version = "0.1.0"
description = "arXiv user admin API"
authors = ["Naoyuki Tai <ntai@arxiv.org>"]
license = "BSD-3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
arxiv-base = {git = "https://github.com/arXiv/arxiv-base.git", rev = "e4d4bfd05e83b04b298f5f67d5a78c41c7ec0ffa"}
pymysql = "^1.1.0"
ruamel-yaml = "^0.18.6"
python-json-logger = "^2.0.7"
python-dateutil = "^2.8.2"
pydantic-sqlalchemy-2 = "^0.0.2"
fastapi-pagination = "^0.12.26"
pyjwt = "^2.9.0"
python-keycloak = "^4.3.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
